requirements.txt
google-api-python-client
google-auth
google-analytics-data
python-dotenv
anthropic
flask
plotly
pandas
schedule
requests
